Add material management module
- Adds module controller for component lifecycle handling
- Adds facade for external module access based on the `IMaterialManagement` facade interface
- Adds the `IContainerPool` component to monitor the known `IMaterialContainer` instances in the system
  - Register added, removed resources
  - Forward update/state-change events from containers
- Adds `IMaterialFlowHandler` component to handle material flow transitions
  - Adds material request handling
  - Adds material pre-advice handling

Also
- Extends resources REST API to retrieve single resource type nodes

+++ src/Tests/Moryx.Material.IntegrationTests/MaterialContainerTests.cs
@@ -32,7 +32,7 @@ internal sealed class MaterialContainerTests : TestBase
         // Arrange
         var hostId = await CreateContainerHostAsync();
         const string material = "Some Material";
-        const decimal quantity = 42m;
+        const double quantity = 42;
         const string unit = "Some Unit";
 
         // Act
@@ -65,7 +65,7 @@ internal sealed class MaterialContainerTests : TestBase
         {
             Id = "Some Id",
             Material = "Some Material",
-            RequestedQuantity = 42m,
+            RequestedQuantity = 42,
             Unit = "Some Unit",
             ExpectedArrival = DateTime.UtcNow,
         };
